+++ src/parser/ast.ts
@@ -4,15 +4,12 @@ import {
   type ASTNodeChildrenSelector,
   type ASTNodeSelector,
   type ASTNodeFirstMatchSelector,
-  defaultASTNodeSelector,
-  defaultASTNodeFirstMatchSelector,
 } from "./selector";
-import type { LazyString } from "../helper";
-import { Lazy } from "../helper";
 import { InvalidTraverseError } from "./error";
 import type { Traverser } from "./traverser";
 import { defaultTraverser } from "./traverser";
 import { anonymousKindPlaceholder } from "../anonymous";
+import type { StringOrLiteral } from "../helper";
 
 /**
  * A structured type for serialization.
@@ -47,50 +44,132 @@ export type ASTObj = {
 };
 
 // TODO: add TraverseContext as a generic type parameter
-export class ASTNode<
-  Kinds extends string, // TODO: rename this to Kind, add NTs as a new type parameter
+export abstract class ASTNode<
+  Kind extends NTs | ExtractKinds<TokenType>, // the kind name of this node
+  NTs extends string, // all NTs' kind names
   ASTData,
   ErrorType,
   TokenType extends GeneralToken,
 > {
   /**
    * T's or NT's kind name.
-   * If the T is anonymous, the value is an empty string.
    */
-  readonly kind: Kinds | ExtractKinds<TokenType>;
+  readonly kind: Kind;
+  /**
+   * By default, this is the same as the kind name.
+   * You can rename nodes in your grammar rules.
+   */
+  name: string;
   /**
    * Start position of the whole input string.
    * Same as the first token's start position.
    */
   readonly start: number;
   /**
-   * T's text content.
-   */
-  readonly text?: string;
-  /**
-   * NT's children.
-   */
-  children?: readonly ASTNode<Kinds, ASTData, ErrorType, TokenType>[]; // TODO: conditional make this non nullable
-  /**
-   * Parent must be an NT node, or `undefined` if this node is a top level node.
-   * This is not readonly because it will be set by parent node.
-   */
-  parent?: ASTNode<Kinds, ASTData, ErrorType, TokenType>; // TODO: conditional make this non nullable
-  /**
-   * Data calculated by traverser.
-   * You can also set this field manually if you don't use top-down traverse.
+   * @default undefined
    */
   data?: ASTData;
-  error?: ErrorType;
   /**
-   * If this is a T, this field is set.
+   * @default undefined
    */
-  token?: TokenType; // TODO: conditional make this non nullable
+  error?: ErrorType;
+
+  // should only be used by subclasses
+  protected constructor(
+    p: Pick<
+      ASTNode<Kind, NTs, ASTData, ErrorType, TokenType>,
+      "kind" | "start" | "data" | "error"
+    >,
+  ) {
+    // name will be set when the parent node is created
+    // so in constructor, we should use the default name
+    this.name = p.kind;
+
+    this.kind = p.kind;
+    this.start = p.start;
+    this.data = p.data;
+    this.error = p.error;
+  }
+
+  /**
+   * Return a tree-structured string.
+   */
+  toTreeString(options?: {
+    /**
+     * The indent for each level.
+     * @default '  ' // two spaces
+     */
+    indent?: string;
+    /**
+     * Anonymous kind name placeholder.
+     * @default "<anonymous>" // anonymousKindPlaceholder
+     */
+    anonymous?: string;
+  }): string {
+    return this._toTreeString({
+      initial: "",
+      indent: options?.indent ?? "  ",
+      anonymous: options?.anonymous ?? anonymousKindPlaceholder,
+    });
+  }
+
+  /**
+   * `toTreeString` with full options.
+   */
+  abstract _toTreeString(options: {
+    /**
+     * The initial indent.
+     */
+    initial: string;
+    /**
+     * The indent for each level.
+     */
+    indent: string;
+    /**
+     * Anonymous kind name placeholder.
+     */
+    anonymous: string;
+  }): string;
+
+  /**
+   * For debug output.
+   */
+  abstract toString(): string;
+
+  /**
+   * Return an ASTObj for serialization.
+   */
+  abstract toJSON(): ASTObj;
+
+  /**
+   * A type guard for checking the kind of this node.
+   */
+  is<TargetKind extends Kind>(
+    kind: TargetKind,
+  ): this is ASTNode<TargetKind, NTs, ASTData, ErrorType, TokenType> {
+    return this.kind === kind;
+  }
+}
+
+export class NTNode<
+  Kind extends NTs, // the kind name of this node
+  NTs extends string, // all NTs' kind names
+  ASTData,
+  ErrorType,
+  TokenType extends GeneralToken,
+> extends ASTNode<Kind, NTs, ASTData, ErrorType, TokenType> {
+  readonly children: readonly ASTNode<
+    NTs | ExtractKinds<TokenType>,
+    NTs,
+    ASTData,
+    ErrorType,
+    TokenType
+  >[];
   /**
    * Select the first matched child node by the name.
    */
   readonly $: ASTNodeFirstMatchChildSelector<
-    Kinds,
+    NTs,
     ASTData,
     ErrorType,
     TokenType
@@ -98,175 +177,99 @@ export class ASTNode<
   /**
    * Select children nodes by the name.
    */
-  readonly $$: ASTNodeChildrenSelector<Kinds, ASTData, ErrorType, TokenType>;
+  readonly $$: ASTNodeChildrenSelector<NTs, ASTData, ErrorType, TokenType>;
+
   /**
    * `traverser` shouldn't be exposed
    * because we want users to use `traverse` instead of `traverser` directly.
    * Make this private to prevent users from using it by mistake.
+   *
+   * Only NT node has traverser, since T node is created by token.
    */
-  private traverser: Traverser<Kinds, ASTData, ErrorType, TokenType>;
-  /**
-   * `name` is set by parent node, so it should NOT be readonly, but can only be set privately.
-   */
-  private _name: string;
-
-  /**
-   * @see {@link ASTNode.toString}
-   */
-  readonly str: LazyString;
-  /**
-   * @see {@link ASTNode.getStrWithName}
-   */
-  readonly strWithName: LazyString;
-  /**
-   * @see {@link ASTNode.getStrWithoutName}
-   */
-  readonly strWithoutName: LazyString;
+  private traverser: Traverser<Kind, ASTData, ErrorType, TokenType>;
 
   constructor(
     p: Pick<
-      ASTNode<Kinds, ASTData, ErrorType, TokenType>,
-      "kind" | "start" | "text" | "children" | "parent" | "data" | "error"
+      NTNode<Kind, NTs, ASTData, ErrorType, TokenType>,
+      "kind" | "start" | "data" | "error" | "children"
     > & {
-      traverser?: Traverser<Kinds, ASTData, ErrorType, TokenType>;
-      selector?: ASTNodeSelector<Kinds, ASTData, ErrorType, TokenType>;
-      firstMatchSelector?: ASTNodeFirstMatchSelector<
-        Kinds,
+      traverser?: Traverser<Kind, ASTData, ErrorType, TokenType>;
+      firstMatchSelector: ASTNodeFirstMatchSelector<
+        NTs,
         ASTData,
         ErrorType,
         TokenType
       >;
-    } & Partial<
-        Pick<ASTNode<Kinds, ASTData, ErrorType, TokenType>, "name" | "token">
-      >,
+      selector: ASTNodeSelector<NTs, ASTData, ErrorType, TokenType>;
+    },
   ) {
-    this._name = p.name ?? p.kind;
-    this.kind = p.kind;
-    this.start = p.start;
-    this.text = p.text;
+    super(p);
     this.children = p.children;
-    this.parent = p.parent;
-    this.data = p.data;
-    this.error = p.error;
-    this.token = p.token;
     this.traverser = p.traverser ?? defaultTraverser;
-    const selector = p.selector ?? defaultASTNodeSelector;
-    const firstMatchSelector =
-      p.firstMatchSelector ?? defaultASTNodeFirstMatchSelector;
-    this.$ = (name: string) => firstMatchSelector(name, this.children ?? []);
-    this.$$ = (name: string) => selector(name, this.children ?? []);
-
-    this.str = new Lazy(
-      () =>
-        `ASTNode({ kind: "${this.kind}", start: ${
-          this.start
-        }, text: ${JSON.stringify(this.text)}, data: ${JSON.stringify(
-          this.data,
-        )}, error: ${JSON.stringify(this.error)} })`,
-    );
-    this.strWithName = new Lazy(() => ASTNode.getStrWithName(this));
-    this.strWithoutName = new Lazy(() => ASTNode.getStrWithoutName(this));
+    this.$ = <
+      TargetKind extends StringOrLiteral<NTs | ExtractKinds<TokenType>>,
+    >(
+      name: TargetKind,
+    ) => p.firstMatchSelector(name, this.children);
+    this.$$ = <
+      TargetKind extends StringOrLiteral<NTs | ExtractKinds<TokenType>>,
+    >(
+      name: TargetKind,
+    ) => p.selector(name, this.children);
   }
 
-  static from<
-    Kinds extends string,
-    ASTData,
-    ErrorType,
-    TokenType extends GeneralToken,
-  >(t: Readonly<TokenType>) {
-    return new ASTNode<Kinds, ASTData, ErrorType, TokenType>({
-      kind: t.kind,
-      start: t.start,
-      text: t.content,
-      token: t,
-    });
+  toString(): string {
+    return `ASTNode(${JSON.stringify({
+      name: this.name,
+      kind: this.kind,
+      start: this.start,
+      children: this.children.map((c) => c.name),
+      // data and error might not able to be serialized
+    })})`;
   }
 
-  /**
-   * By default, this is the same as the kind name.
-   * You can rename nodes in your grammar rules.
-   */
-  get name() {
-    return this._name;
-  }
+  // _toTreeString(indent: string): string {
+  //   const kind = this.kind === "" ? anonymousKindPlaceholder : this.kind;
+  //   let res = `${indent}${
+  //     this.kind === this.name ? kind : `${kind}@${this.name}`
+  //   }: `;
+  //   if (this.text) res += JSON.stringify(this.text); // quote the text
+  //   res += "\n";
+  //   this.children?.forEach((c) => {
+  //     res += c.toTreeString({
+  //       indent: indent + "  ",
+  //     });
+  //   });
+  //   return res;
+  // }
 
-  set name(name: string) {
-    this._name = name;
-    this.strWithName.reset(); // clear the cache
-  }
+  _toTreeString(options: {
+    initial: string;
+    indent: string;
+    anonymous: string;
+  }): string {
+    const kind = this.kind === "" ? options.anonymous : this.kind;
 
-  /**
-   * Return a tree-structured string.
-   * The result is NOT cached.
-   */
-  toTreeString(options?: { indent?: string }) {
-    const indent = options?.indent ?? "";
-
-    // don't use `this.toStringWithName` here
-    // since this output will always have ': '
-    const kind = this.kind === "" ? anonymousKindPlaceholder : this.kind;
-    let res = `${indent}${
+    return `${options.initial}${
       this.kind === this.name ? kind : `${kind}@${this.name}`
-    }: `;
-    if (this.text) res += JSON.stringify(this.text); // quote the text
-    res += "\n";
-    this.children?.forEach((c) => {
-      res += c.toTreeString({
-        indent: indent + "  ",
-      });
-    });
-    return res;
+    }: \n${this.children
+      .map((c) =>
+        c._toTreeString({
+          initial: options.initial + options.indent,
+          indent: options.indent,
+          anonymous: options.anonymous,
+        }),
+      )
+      .join("")}`;
   }
 
-  /**
-   * For debug output.
-   */
-  toString() {
-    return this.str.value;
-  }
-
-  /**
-   * Format: `kind@name: text`.
-   * This value will be changed if you change the name of this node.
-   */
-  static getStrWithName(
-    data: Pick<
-      ASTNode<string, unknown, unknown, GeneralToken>,
-      "kind" | "name" | "text"
-    >,
-  ) {
-    return (
-      `${data.kind === "" ? anonymousKindPlaceholder : data.kind}` +
-      `${data.name === data.kind ? "" : `@${data.name}`}` +
-      `${data.text === undefined ? "" : `: ${JSON.stringify(data.text)}`}`
-    );
-  }
-
-  /**
-   * Format: `kind: text`.
-   */
-  static getStrWithoutName(
-    data: Pick<
-      ASTNode<string, unknown, unknown, GeneralToken>,
-      "kind" | "text"
-    >,
-  ) {
-    return (
-      `${data.kind === "" ? anonymousKindPlaceholder : data.kind}` +
-      `${data.text === undefined ? "" : `: ${JSON.stringify(data.text)}`}`
-    );
-  }
-
-  /**
-   * Return an ASTObj for serialization.
-   */
   toJSON(): ASTObj {
     return {
       name: this.name,
       kind: this.kind,
       start: this.start,
-      text: this.text || "",
-      children: this.children?.map((c) => c.toJSON()) ?? [],
+      text: "",
+      children: this.children.map((c) => c.toJSON()),
     };
   }
 
@@ -279,7 +282,7 @@ export class ASTNode<
   traverse(): ASTData | undefined {
     if (this.children === undefined) throw new InvalidTraverseError(this);
     const res = this.traverser(
-      this as Parameters<Traverser<Kinds, ASTData, ErrorType, TokenType>>[0], // children is not undefined
+      this as Parameters<Traverser<Kind, ASTData, ErrorType, TokenType>>[0], // children is not undefined
     );
     this.data =
       res ??
@@ -288,5 +291,89 @@ export class ASTNode<
         : // undefined or void
           undefined);
     return this.data;
+  }
+}
+
+export class TNode<
+  Kind extends ExtractKinds<TokenType>, // the kind name of this node
+  NTs extends string, // all NTs' kind names
+  ASTData,
+  ErrorType,
+  TokenType extends GeneralToken,
+> extends ASTNode<Kind, NTs, ASTData, ErrorType, TokenType> {
+  /**
+   * Parent must be an NT node, or `undefined` if this node is a top level node.
+   * This is not readonly because it will be set by parent node.
+   */
+  parent?: NTNode<NTs, NTs, ASTData, ErrorType, TokenType>;
+  token: TokenType & { kind: Kind };
+
+  /**
+   * Token's text content.
+   */
+  get text() {
+    return this.token.content;
+  }
+
+  constructor(
+    p: Pick<
+      TNode<Kind, NTs, ASTData, ErrorType, TokenType>,
+      "kind" | "start" | "data" | "error" | "parent" | "token"
+    >,
+  ) {
+    super(p);
+    this.parent = p.parent;
+    this.token = p.token;
+  }
+
+  static from<
+    NTs extends string,
+    ASTData,
+    ErrorType,
+    TokenType extends GeneralToken,
+  >(t: Readonly<TokenType>) {
+    return new TNode<
+      ExtractKinds<TokenType>,
+      NTs,
+      ASTData,
+      ErrorType,
+      TokenType
+    >({
+      kind: t.kind,
+      start: t.start,
+      token: t,
+    });
+  }
+
+  toString(): string {
+    return `ASTNode(${JSON.stringify({
+      name: this.name,
+      kind: this.kind,
+      start: this.start,
+      text: this.text,
+      // data and error might not able to be serialized
+    })})`;
+  }
+
+  _toTreeString(options: {
+    initial: string;
+    indent: string;
+    anonymous: string;
+  }): string {
+    const kind = this.kind.length === 0 ? options.anonymous : this.kind;
+
+    return `${options.initial}${
+      this.kind === this.name ? kind : `${kind}@${this.name}`
+    }: ${JSON.stringify(this.text)}\n`;
+  }
+
+  toJSON(): ASTObj {
+    return {
+      name: this.name,
+      kind: this.kind,
+      start: this.start,
+      text: this.text,
+      children: [],
+    };
   }
 }